Added Telemetry to Commands
Updated the COR interfaces to be more GoLang centric, now requiring an
outer context to be passed to the execution blocks. You can see this in
the IsExecutable() method.

Updated all commands and chains to use the context accordingly. This
ensures that all cloud interactsions can participate in the outer most
trace.

Updated the build scripts and dependencies to pull in the open telemetry
stack.

Verified that all spans are working in accordance to the outermose
transaction id and added baggage around the gemini calls to further
understand where failures are occuring.

Added "names" to all commands to aid in the contextualization of
execution.

Added names to chains for scope clarity.